product/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
psycopg2-binary
redis
python-dotenv
influxdb-client
gunicorn